Refactor: Combined Adapters packages
