+++ app/src/main/java/com/app/venki/up/adapters/GroupOnRecycler.java
@@ -1,4 +1,8 @@
-package com.app.venki.up.recycleradapters;
+package com.app.venki.up.adapters;
+
+/**
+ * Created by yash on 8/20/2016.
+ */
 
 import android.content.Context;
 import android.support.v7.widget.CardView;
@@ -10,18 +14,18 @@ import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
 import com.app.venki.up.R;
-import com.app.venki.up.model.coupons.CouponsArray;
+import com.app.venki.up.model.coupons.GroupOnEvent;
 
-/**
- * Created by Billy on 7/9/16.
- */
-public class CouponsRecyclerAdapter extends android.support.v7.widget.RecyclerView.Adapter<CouponsRecyclerAdapter.RecyclerViewHolder> {
 
-    private CouponsArray data;
+public class GroupOnRecycler extends android.support.v7.widget.RecyclerView.Adapter<GroupOnRecycler.RecyclerViewHolder> {
+
+    private GroupOnEvent data;
     private Context context;
-    private CouponClickListener couponClickListener;
+    private CouponClickListener2 couponClickListener;
 
-    public interface CouponClickListener{
+
+
+    public interface CouponClickListener2{
         void onCardViewClick(String link);
     }
 
@@ -41,7 +45,7 @@ public class CouponsRecyclerAdapter extends android.support.v7.widget.RecyclerVi
             info = (TextView) itemView.findViewById(R.id.info_ID);
         }
 
-        public void bind(final CouponClickListener couponClickListener, final String link){
+        public void bind(final CouponClickListener2 couponClickListener, final String link){
             itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View view) {
@@ -52,21 +56,18 @@ public class CouponsRecyclerAdapter extends android.support.v7.widget.RecyclerVi
     }
 
 
-    public CouponsRecyclerAdapter(CouponClickListener couponClickListener, CouponsArray data) {
-        this.couponClickListener = couponClickListener;
+    public GroupOnRecycler(CouponClickListener2 couponClickListener, GroupOnEvent data) {
+        this.couponClickListener =  couponClickListener;
         this.data = data;
     }
 
-    @Override
     public void onBindViewHolder(RecyclerViewHolder holder, int position) {
 
-        holder.title.setText(data.getCoupons().get(position).getDealTitle());
-        holder.info.setVisibility(View.GONE);
 
-        String url = data.getCoupons().get(position).getURL();
+        holder.title.setText(data.getResults().get(position).getTitle());
+        String url =data.getResults().get(position).getDealUrl();
         holder.bind(couponClickListener, url);
-
-        String imageURI = data.getCoupons().get(position).getShowImageStandardBig();
+        String imageURI =data.getResults().get(position).getMediumImageUrl();//testing
 
         if (imageURI.isEmpty()) {
             imageURI = "R.drawable.blank_white.png";
@@ -90,7 +91,7 @@ public class CouponsRecyclerAdapter extends android.support.v7.widget.RecyclerVi
         context = parent.getContext();
         LayoutInflater inflater = LayoutInflater.from(context);
 
-        View view = inflater.inflate(R.layout.recyclerview_custom_layout, parent, false);
+        View view = inflater.inflate(R.layout.cardview_coupons, parent, false);
         RecyclerViewHolder vh = new RecyclerViewHolder(view);
 
         return vh;
@@ -98,8 +99,6 @@ public class CouponsRecyclerAdapter extends android.support.v7.widget.RecyclerVi
 
     @Override
     public int getItemCount() {
-        return data.getCoupons().size();
+        return data.getResults().size();//testing
     }
-
-
 }
